remove tests folder to remove any duplicates
